First commit post fork of quill
sns-quill will now be a stand-alone tool for interacting with the SNS.
This first commit "guts" the files not needed or not compatible. There
will be followup commits for adding back supported commands. The
libraries under src/lib have been updated with the required changes
to support the SNS. Additionally, the dependencies have been
updated to include sns common types, candid 0.7.14, ic-agent 0.15.0,
and ic-types 0.3.0.