Fix `quill generate` arg (#131)
A bug was introduced with `quill generate` in #126 around the repetition of a `clap` argument. Because the `pem_file` and `seed_file` parameters were repeated twice in the same command, running the command would crash. This fixes it by removing the common flags from the command, as they are not used at all.