Remove merging maturity, add staking maturity (#140)
Per SDK-772, merging maturity is being deprecated and replaced with staking maturity (maturity staked as a separate asset factored into voting power as opposed to being converted to ICP and added to the primary stake). This adds the new maturity staking function to `quill neuron-manage`, as well as converting the existing maturity merging operation to an error describing the deprecation.